add integration test for java-library plugin
